move lockfile parser tests to their own file
